refactor: Client gets a file for each invocation
