+++ pkg/client/spaceblobadd_test.go
@@ -93,7 +93,7 @@ func TestBlobAdd(t *testing.T) {
 
 	testBlob := bytes.NewReader([]byte("test"))
 
-	_, _, err = client.BlobAdd(context.Background(), testBlob, issuer, space.DID(), receiptsURL, client.WithConnection(conn), client.WithProof(proof))
+	_, _, err = client.SpaceBlobAdd(context.Background(), testBlob, issuer, space.DID(), receiptsURL, client.WithConnection(conn), client.WithProof(proof))
 	require.NoError(t, err)
 }
